Ref: moving isPrimitive() to zts helpers.
